Implement parallel test Runner with detailed reporting
This adds the Runner class which executes the tests in parallel.

Number of workers defaults to the number of cores on the system but
can be changed with -c.

The Runner features a real-time progress bar to indicate its progress.

The Runner collects all the outcomes in a nested dict and reports them
summed up for each directory in the tests tree. Directories with 0 tests
passed are not expanded.