CMake: Set BUILD_LAGOM as an internal cache variable
We always want to BUILD_LAGOM. This regressed in serenity commit
https://github.com/SerenityOS/serenity/commit/904a268872416dc9b9f26c83fd0b1bd8b715c511
where BUILD_LAGOM began being set as a cache variable. The impact of
this on the libjs-test262 build is that after BUILD_LAGOM was set to OFF
in the cache, it would be overriden by the normal variable in FetchLagom
but not the first time it gets ran.