libjs-test262-runner: Add "todo_error" result
If an InternalError with its message starting with "TODO(" is thrown we
assume we have hit a TODO exception.
We also check for a special case where the InternalError was caught when
another error was expected. This works because we currently only use
InternalError for these TODO parts but in the future might give false
positives. However these false positives still indicate a failing tests
so they are only slightly deceiving and should not influence the actual
passing tests.