libjs-test262-runner: Add a mode for forwarding all stderr output
It simply outputs all the stderr output from the process into the stderr
of the python script. Since the progress bar is also displayed on the
stderr output we have to use tdqm methods to make sure we don't
duplicate the progress bar.

Unfortunately if you want to capture this output into a file you might
want to enable silent mode else the progress bar also appears
in the output possible mixed in the messages.